v1.0: Basic Features
Opens the camera feed and starts acquiring when the video is available.
The video is put in the background as a canvas.

On clicking anywhere on the window, the X and Y coordinates are
registered and saved into a table file.

Program can be stopped at the press of ENTER key.