New Version written in Python
Processing does not allow creating a transparent canvas which can be
overlayed on top of a camera feed in a 3rd party software. But it does
have a capability to directly take the camera feed and make it a
background of the canvas. This can work if the camera is detected and is
readable by processing. I could not make this step work. Also,
overlaying the window on the camera feed appears to be an easier, yet
not versatile solution.

Therefore, i wrote this new version in python using Tkinter library
which allows GUI and transparency. The script currently produces a fixed
sized transparent window which can be overlayed on the camera feed. The
clicks on the window will be registered and their coordinates scaled to
the polygon frame size will be saved as a csv file.

What next: The next step is to create a csv file with the coordinates of
the pixels which will form a circle arround the clicked location.